+++ Examples/Qen.cs
@@ -49,7 +49,7 @@ namespace PLGL.Examples
         private void SetFilters()
         {
             lang.AddFilter("Delimiter", " ");
-            lang.AddFilter("Compound", "");
+            lang.AddFilter("Compound", "");//-
 
             lang.AddFilter("Letters", "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ");
             lang.AddFilter("Numbers", "1234567890");
@@ -155,21 +155,42 @@ namespace PLGL.Examples
             lang.Structure.AddGroup('P', "Plosive", ('b', 3.0), ('p', 4.0), ('d', 7.0), ('t', 1.0), ('g', 6.0), ('k', 2.0));
             lang.Structure.AddGroup('p', "Plosive higher", ('p', 4.0), ('t', 1.0), ('k', 2.0));
             lang.Structure.AddGroup('F', "Fricative", ('f', 1.0), ('v', 1.0), ('s', 1.0), ('z', 1.0));
-            lang.Structure.AddGroup('S', "Fricative common", ('s', 30.0), ('ŝ', 1.0));
+            lang.Structure.AddGroup('S', "S/SH", ('s', 30.0), ('ŝ', 1.0));
             lang.Structure.AddGroup('A', "Approximant", ('w', 1.0), ('y', 1.0), ('h', 1.0));
             lang.Structure.AddGroup('R', "R/L", ('r', 66), ('l', 33));
+            lang.Structure.AddGroup('T', "TH/SH", ('Þ', 1.0), ('ŝ', 1.0));
 
-            lang.Structure.AddSyllable("VP", 1.0);
+            lang.Structure.AddSyllable("VP", 0.5);
             lang.Structure.AddSyllable("NVP", 2.0);
             lang.Structure.AddSyllable("SpVN", 3.0);
+            lang.Structure.AddSyllable("SpRVP", 0.5);
+            lang.Structure.AddSyllable("SpRVN", 1.0);
+
+            lang.Structure.AddSyllable("TVN", 0.75);
+            lang.Structure.AddSyllable("TVR", 0.75);
+            lang.Structure.AddSyllable("TVn", 0.25);
+
+
+            lang.Structure.AddSyllable("VS", 0.25);
 
             lang.Structure.AddSyllable("PV", 1.0);
             lang.Structure.AddSyllable("PVR", 5.0);
             lang.Structure.AddSyllable("PVN", 2.0);
 
             lang.Structure.AddSyllable("VN", 1.0);
-            lang.Structure.AddSyllable("AV", 1.0);
+            lang.Structure.AddSyllable("AVR", 1.0);
             lang.Structure.AddSyllable("on", 0.5);
+
+            lang.OnLetter += (lg, word, letter) =>
+                lg.LETTER_Replace(word, letter, letter.AdjacentLeft, 'ü', lg.LETTER_Contains(letter.AdjacentLeft, 'u'));
+
+            //Manual consonant doubling, depending on syllable condition—group type, letter type ('l'), syllable location.
+            //(where "T" is th/sh letters, "V" are vowels, and "R" is r or l.
+            /*lang.OnLetter += (lg, word, letter) =>
+                lg.LETTER_Insert(word, letter, 'l', 0,
+                lg.LETTER_Syllable(letter, "TVR") &&
+                lg.LETTER_Contains(letter, 'l') &&
+                (letter.Syllable.SyllableIndex < word.Syllables.Count - 1));*/
         }
         #endregion